# 🌍 #349 Add bespoke Zip implementation, improves performance by 25% over SelectMany derivation