--- a/src/GalaxyCheck/ExampleSpaces/ExampleId.cs
+++ b/src/GalaxyCheck/ExampleSpaces/ExampleId.cs
@@ -1,18 +1,50 @@
 ﻿namespace GalaxyCheck.ExampleSpaces
 {
-    public record ExampleId(int HashCode)
+    public abstract record ExampleId
     {
-        public static ExampleId Empty => new ExampleId(-1923861349);
+        private ExampleId() { }
+
+        private record CaseEmpty() : ExampleId()
+        {
+            public override string ToString()
+            {
+                return "<empty>";
+            }
+        }
+
+        private record CasePrimitive(long HashCode) : ExampleId()
+        {
+            public override string ToString()
+            {
+                return HashCode.ToString();
+            }
+        }
+
+        public static ExampleId Empty => new CaseEmpty();
+
+        public static ExampleId Primitive(long id) => new CasePrimitive(id);
 
         public static ExampleId Primitive(object? obj) => obj == null
             ? Empty
-            : new ExampleId(obj.GetHashCode());
+            : new CasePrimitive(obj.GetHashCode());
 
         public static ExampleId Combine(ExampleId left, ExampleId right)
         {
-            if (left == Empty) return right;
-            if (right == Empty) return left;
-            return Primitive(left.HashCode * -1521134295 + right.HashCode);
+            return (left, right) switch
+            {
+                (CaseEmpty, var r) => r,
+                (var l, CaseEmpty) => l,
+                (CasePrimitive l, CasePrimitive r) => CombinePrimitives(l, r),
+                _ => throw new System.Exception("Unhandled case")
+            };
+        }
+
+        private static ExampleId CombinePrimitives(CasePrimitive l, CasePrimitive r)
+        {
+            unchecked
+            {
+                return Primitive(l.HashCode * -1521134295 + r.HashCode);
+            }
         }
     }
 }