SelectMany forces the gens to be created, as they are derived from the values of the previous gens. Zip doesn't need this - the values from the zipped generators are not dependent on one another.

Also moved the ExampleId into long-space - new Zip revealed some cases where int-space wasn't good enough (as I expected may have been the case, way back when).

Before:

```
|                         Method | iterations |         Mean |       Error |       StdDev |       Median |
|------------------------------- |----------- |-------------:|------------:|-------------:|-------------:|
|    SampleClassWith10IntMembers |          1 |     664.7 us |    29.24 us |     83.89 us |     632.7 us |
| SampleClassWith10StringMembers |          1 |  10,776.6 us | 1,237.96 us |  3,630.73 us |   9,971.3 us |
|    SampleClassWith10IntMembers |         10 |   3,203.2 us |   261.19 us |    736.68 us |   2,993.1 us |
| SampleClassWith10StringMembers |         10 |  40,563.0 us | 1,094.87 us |  3,123.73 us |  40,131.2 us |
|    SampleClassWith10IntMembers |        100 |  25,980.5 us | 1,464.98 us |  4,296.53 us |  24,586.4 us |
| SampleClassWith10StringMembers |        100 | 493,894.1 us | 9,622.21 us | 16,339.24 us | 491,388.3 us |
```

After:

```
|                         Method | iterations |         Mean |       Error |       StdDev |       Median |
|------------------------------- |----------- |-------------:|------------:|-------------:|-------------:|
|    SampleClassWith10IntMembers |          1 |     566.1 us |    23.15 us |     67.16 us |     534.4 us |
| SampleClassWith10StringMembers |          1 |   8,482.0 us |   781.16 us |  2,278.67 us |   8,443.4 us |
|    SampleClassWith10IntMembers |         10 |   2,155.7 us |    70.04 us |    204.32 us |   2,118.2 us |
| SampleClassWith10StringMembers |         10 |  35,987.0 us | 1,074.72 us |  2,995.88 us |  35,888.9 us |
|    SampleClassWith10IntMembers |        100 |  15,725.4 us |   837.40 us |  2,442.74 us |  14,857.3 us |
| SampleClassWith10StringMembers |        100 | 399,173.6 us | 7,848.37 us | 16,554.88 us | 401,153.7 us |
```